+++ app/server/server_db.R
@@ -4,7 +4,8 @@
 rv <- reactiveValues(
   db_avail = FALSE,
   db_data = NULL,
-  new_data = NULL,
+  val_import = NULL,
+  import_dat = NULL,
   clear = FALSE
 )
 
@@ -24,10 +25,9 @@ output$db_present <- reactive({
 
 outputOptions(output, "db_present", suspendWhenHidden=FALSE)
 
-
 # Render database ----
 output$db_files <- renderDataTable(rv$db_data %>% 
-                                     select(date, exp_name) %>% 
+                                     select(date, exp_name, instrument, id) %>% 
                                      mutate(date = as.character(date)) %>%
                                      arrange(date) %>%
                                      distinct(),
@@ -39,34 +39,31 @@ observe({
   req(input$input_files)
   req(!rv$clear)
   
-  file_paths <<- input$input_files # TODO: replace <<- --> <-?
+  file_paths <- input$input_files
+  file_paths <- validateFiles(file_paths)
   
-  # Preprocess files df
-  # print(file_paths)
-  file_paths <<- file_paths %>%
-    rename(exp_name = name) %>% 
-    mutate(exp_name = substr(exp_name, 0, nchar(exp_name)-4))
-  
-  # TODO: Validate files to prompt error message
-  # mes <- validateFileDF(file_paths)
-  
-  # For now, import and preprocess data
-  file_paths <<- file_paths %>% as_tibble() %>% 
-    mutate(map_df(datapath, readTxt), 
-           data = map(data, processQSResults))
+  n_files_for_import <- ifelse(all(file_paths$error == ""), nrow(file_paths), 0)
   
   # Check if db is available in order to draw table
   if(is.null(rv$db_data)) {
-    file_paths <<- file_paths %>% mutate(in_database = "No")
+    file_paths <- file_paths %>% mutate(in_database = "No")
   } else {
     # Check if not available
-    file_paths <<- file_paths %>% 
+    file_paths <- file_paths %>% 
       mutate(in_database = if_else(exp_name %in% rv$db_data$exp_name, "Yes", "No"))
   }
-    
-  rv$new_data <- file_paths %>%
-    select(date, exp_name, in_database) %>%
+  # TODO: move the "in_database" info to "error", and rename 'error' to message
+  
+  rv$import_dat <- file_paths %>% 
+    select(-error)
+  
+  rv$val_import <- file_paths %>% 
+    select(date, exp_name, in_database, error) %>%
     mutate(date = as.character(date))
+
+  # Update button
+  updateActionButton(session, "add_to_db",
+                     label = paste("Import ", n_files_for_import, " files"))
 }
 )
 
@@ -76,38 +73,52 @@ observeEvent(input$input_files, {
 
 # Check whether files are imported ----
 output$files_imported <- reactive({
-  return(!is.null(rv$new_data))
+  return(!is.null(rv$val_import))
 })
 
 outputOptions(output, "files_imported", suspendWhenHidden=FALSE)
 
 # Render new files ----
-output$input_files_table <- renderDataTable(rv$new_data, options = list(pageLength = 10))
+output$input_files_table <- renderDataTable(rv$val_import, 
+                                            options = list(pageLength = 10))
+
+# Check whether there are no errors ----
+output$no_errors <- reactive({
+  return(all(rv$val_import$error == "") & all(rv$val_import$in_database == "No"))
+})
+
+outputOptions(output, "no_errors", suspendWhenHidden=FALSE)
 
 # Add files to database and reset ----
 observeEvent(input$add_to_db, {
   # Check for presence database
   if(!is.null(rv$db_data)) {
     # Prepare x by binding the current db to the added samples
-    if(sum(file_paths$in_database == "No") > 0) {
-      db <- rv$db_data %>% 
-        bind_rows(file_paths %>% 
-                    filter(in_database == "No") %>% 
-                    unnest(data) %>% 
-                    select(date, exp_name, instrument, id, target, ct)) %>% 
-                    arrange(date)  
+    if(sum(rv$import_dat$in_database == "No") > 0) {
+      db <- rv$db_data %>%
+        bind_rows(rv$import_dat %>%
+                    filter(in_database == "No") %>%
+                    unnest(data) %>%
+                    select(date, exp_name, instrument, id, target, ct)) %>%
+                    arrange(date)
     } else {
       db <- rv$db_data
     }
+  } else {
+    db <- rv$import_dat %>%
+      filter(in_database == "No") %>%
+      unnest(data) %>%
+      select(date, exp_name, instrument, id, target, ct) %>%
+      arrange(date)
   }
-  
+
   rv$db_data <- db
-  
+
   # Prepare and save x
   write_csv(x = db, file = "./../db.csv")
-  
+
   # Reset infiles and table
-  rv$new_data <- NULL
+  rv$val_import <- NULL
   rv$clear <- TRUE
   rv$db_avail <- TRUE
   reset("input_files")